responsive news page

+++ admin/includes/view_all_news.php
@@ -9,13 +9,12 @@
 ?>
 
 <!-- DISPLAY NEWS ARTICLES ON ADMIN PAGE -->
-<table class="table-bordered table-hover">
+<table class="table table-responsive table-bordered table-hover">
     <thead>
         <tr>
-            <th>Id</th>
             <th>Title</th>
             <th>Date</th>
-            <th>Image</th>
+            <th>Image</th%>
             <th>Content</th>
             <th>Tags</th>
             <th>Source Link</th>
@@ -27,13 +26,15 @@
     <tbody>
         <?php 
         
-        $query = "SELECT * FROM news";
+        $query = "SELECT * FROM news ORDER BY post_date DESC";
         $select_news = mysqli_query($connection, $query);
 
         while ($row = mysqli_fetch_assoc($select_news)) {
+           
             $post_id = $row['post_id'];
             $post_title = $row['post_title'];
             $post_date = $row['post_date'];
+            $formated_date = date('d-m-Y',strtotime($post_date));;
             $post_image = $row['post_image'];
             $post_content = $row['post_content'];
             $post_tags = $row['post_tags'];
@@ -41,12 +42,11 @@
             $source_link_name = $row['source_link_name'];
 
             echo "<tr>";
-            echo "<td>{$post_id}</td>";
             echo "<td>{$post_title}</td>";
-            echo "<td>{$post_date}</td>";
+            echo "<td>{$formated_date}</td>";
             echo "<td>";
             if($post_image != "") {
-                echo "<img width='100' src='../img/{$post_image}' alt='image'>";
+                echo "<img width='100px' src='../img/{$post_image}' alt='image'>";
             }
             echo "</td>";
             echo "<td>{$post_content}</td>";
@@ -54,7 +54,7 @@
             echo "<td>{$source_link}</td>";
             echo "<td>{$source_link_name}</td>";
             echo "<td><a href='news.php?source=edit_news&p_id={$post_id}'>Edit</a></td>";
-            echo "<td><a href='news.php?delete={$post_id}'>Delete</a></td>";
+            echo "<td><a href='news.php?delete={$post_id}' onClick=\"javascript:return confirm('Delete {$post_title}?');\">Delete</a></td>";
             echo "</tr>";
         }                 
         ?>
